+++ agate/warns.py
@@ -17,3 +17,18 @@ def warn_null_calculation(operation, column):
         column.name,
         operation.__class__.__name__
     ), NullCalculationWarning)
+
+
+class DuplicateColumnWarning(RuntimeWarning):  # pragma: no cover
+    """
+    Warning raised if multiple columns with the same name are added to a new
+    :class:`.Table`.
+    """
+    pass
+
+
+def warn_duplicate_column(column_name, column_rename):
+    warnings.warn('Column name "%s" already exists in Table. Column will be renamed to "%s".' % (
+        column_name,
+        column_rename
+    ), DuplicateColumnWarning)